Remove 'started' field from prompt presenter
Some investigation revealed that view state is not restored until
relatively late in the Activity lifecycle in onRestoreInstanceState,
which is called between onStart and onPostCreate. See:

http://developer.android.com/reference/android/app/Activity.html#onRestoreInstanceState(android.os.Bundle)

Because of this, we cannot use the view state restoration cycle to
determine whether a prompt presenter is mid-flow or not when
promptIfReady is called. Instead, we'll need to advise users to only
call promptIfReady inside a check that determines whether the current
activity/fragment is being recreated or not.